+++ CHANGELOG.md
@@ -3,4 +3,5 @@ All notable changes to this project will be documented in this file.
 
 ### [Unreleased]
 ##
-- [0.0.1]: All data extraction from the Accounts part of the API is completed.
+- [0.0.2]  Tankopedia: Maps, Vehicles, Achievements and Information have been completed.
+- [0.0.1]  All data extraction from the Accounts part of the API is completed.